Move remote archived form loading to helper class
